Refactor: 채팅 로직 분리 + GA 적용
