+++ package.json
@@ -13,6 +13,7 @@
     "lodash": "^4.17.21",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-ga4": "^2.1.0",
     "react-icons": "^4.11.0",
     "react-modal": "^3.16.1",
     "react-query": "^3.39.3",
