+++ src/components/SendMessage/SendMessage.tsx
@@ -1,24 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { createMessages } from '../../firebase/createMessages';
-import shortId from 'shortid';
-import getNickname from 'api/getNickname';
 import { RiSendPlaneFill } from 'react-icons/ri';
 import * as styled from './SendMessage.styles';
 import swal from 'sweetalert';
-
-function generateRandomUID() {
-  const randomString = shortId.generate();
-  const uppercaseString = randomString.replace(/[a-z]/g, (char) =>
-    char.toUpperCase(),
-  );
-  return uppercaseString;
-}
-async function saveUserInfoToLocalStorage() {
-  const uid = generateRandomUID();
-  const displayName = await getNickname();
-  localStorage.setItem('uid', uid);
-  localStorage.setItem('displayName', displayName);
-}
 
 const SendMessage = () => {
   const [message, setMessage] = useState('');
@@ -35,14 +19,6 @@ const SendMessage = () => {
     createMessages(message);
     setMessage('');
   };
-
-  useEffect(() => {
-    const storedUid = localStorage.getItem('uid');
-    const storedDisplayName = localStorage.getItem('displayName');
-    if (!storedUid || !storedDisplayName) {
-      saveUserInfoToLocalStorage();
-    }
-  }, []);
 
   return (
     <styled.SendMessageForm onSubmit={(event) => handleSendMessage(event)}>
